Create third interaction design
Create third interaction design to 'HandTracking.cs' file (mode == 3)
Update float numbers (ex: (float)0.2 -> 0.2f)

+++ Assets/ProgressFillShader.cs
@@ -51,7 +51,7 @@ public class ProgressFillShader : MonoBehaviour
         }
         else
         {
-            progressBorder = (float)-0.6;
+            progressBorder = -0.6f;
         }
  
         objectMaterial.SetFloat(name: "_ProgressBorder", progressBorder);
